Tip of the Day (#3118)
* Basic tip of the day with sample widget added

* "Show tips on startup" option added to settings

* tip cycling implemented

* Structure of the tipOfTheDay class and resource created

* tip getter function modified

* Resources added, feature works properly

* clangified

* accidental modification rolled back

* zach cleanup

* tips to spaces; cmake list combined

* cleanup img

* fix copy

* remove TOTD as QObject so we can copy construct it

* prevent mem leaks in dlg

* changed order of 'next' and 'previous' buttons

* Date and tip numbers added; content wraps around

* useless sizepolicy removed

* link support added & clangified

* Initial tips & memory management updates

+++ cockatrice/src/setsmodel.cpp
@@ -220,8 +220,8 @@ void SetsModel::sort(int column, Qt::SortOrder order)
 void SetsModel::save(CardDatabase *db)
 {
     // order
-    for (unsigned int i = 0; i < sets.size(); i++)
-        sets[i]->setSortKey(i + 1);
+    for (int i = 0; i < sets.size(); i++)
+        sets[i]->setSortKey(static_cast<unsigned int>(i + 1));
 
     // enabled sets
     for (const CardSetPtr &set : sets)
